backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
marshmallow==3.19.0
python-dotenv==1.0.0
flask-smorest==0.41.0
pytest==7.3.2
pytest-flask==1.2.0
